"zero shot output"

+++ workflows/baseline_qwen_single_Agent.py
@@ -8,20 +8,12 @@ from tqdm import tqdm
 # ----------------------------
 # LOAD MODEL
 # ----------------------------
-# def load_model(model_path):
-#     tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
-#     model = AutoModelForCausalLM.from_pretrained(
-#         model_path,
-#         torch_dtype=torch.float16,
-#         trust_remote_code=True
-#     ).to("cuda") 
-#     return tokenizer, model
-
 def load_model(model_path):
     tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
     model = AutoModelForCausalLM.from_pretrained(
         model_path,
-        torch_dtype=torch.float16,
+        #torch_dtype=torch.float16,
+        torch_dtype=torch.bfloat16,
         device_map=None,   
         trust_remote_code=True
     )
